add Quadus randomizer
Here is the Quadus randomizer: a chaotic, quirk-filled generator based
on the concept of rebalancing the weight of pieces as pieces are dealt.

For each piece dealt, the dealt piece's weight decreases by one less
than the total types of pieces, while the other pieces' weight increase
by one. Thus, recent, more frequently dealt pieces are less likely to
be dealt, and pieces that have not been seen for a long time are more
likely to be dealt.

An inherent quirk is that the weight can go below zero. When the weight
for a piece goes below zero, not only wouldn't it be dealt, the effect-
ive weight of another piece decreases as well. The order of influence
is determined by the mapping from piece type to numbers; in Quadus the
order is I->J->L->O->S->T->Z.