Move top out checking into ruleset
The only top out check remaining in the game class is checking if
incoming lines cause the stack to exceed 48 lines, colliquially
known as "garbage out".

Also allows initial hold and initial rotation to escape top out.

QUIRK: The O piece technically shares the same kick table with
J, L, S, T and Z in `standard` ruleset. Normally natural rotation
always succeeds with O, but on it can fail when O overlaps with
the stack on spawn, causing later entries of the kick table to
be evaluated. By abusing initial rotation, it is possible for O
to "move out" of its spawn location, escaping a top out.