pull in replacement malloc for some libcs
Esp. MSVCRT, which don't provide an appropriate allocator that supports
aligned_alloc. Solution is to pull in jemalloc, or mimalloc if jemalloc
is not found.

Also fixed several issues so Quadus core can compile for Windows via
mingw-w64; tests are not done yet. No plans to support MSVC; port of
TUI would be done soon(TM).