evdev: null-sink stdin on read
As keyboard is the only currently supported device for evdev input,
inputs would be simultaneously sent to the device file we are reading
and stdin. Unhandled stdin would eventually show up as garbage input
for some shells (e.g. bash) after we exit.

Since reading-on-close does not help with the input issue, stdin is
read right after each read on the evdev device; the data is discarded
right away as everything we want is already read elsewhere.